refactor(Team): refactor TeamMemberResponse, TeamService
