+++ src/main/java/com/example/taskflow/domain/team/controller/TeamController.java
@@ -1,4 +1,5 @@
 package com.example.taskflow.domain.team.controller;
 
 public class TeamController {
+
 }
