+++ src/main/java/com/example/taskflow/domain/team/dto/Request/TeamMemberRequest.java
@@ -4,5 +4,5 @@ import lombok.Getter;
 
 @Getter
 public class TeamMemberRequest {
-    private Long memberId;
+    private Long userId;
 }
